Reworked library system
Uses modules and classes now, with each library using a top-level-await for initialization. These awaits run in opposite order of imports: an imported library will be fully initialized before the one importing it tries to initialize.

Also includes loader.js which is the actual Storyline JS code to include in the project.