Reworked loading system
Now pulls list of modules (and whether or not to use localhost) from Storyline vars, and sets a new Storyline var to true when all modules are loaded.

Actual loader is loader.js, with the JS to include in Storyline in loader.storyline.js (the loader loader, if you will).